+++ packages/CollapsibleCard/CHANGELOG.md
@@ -9,3 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.1.2-alpha.1] - 2020-12-04
 
 - Support render function for `label` prop.
+
+## [1.1.8-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.